# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: VPN Pulse API
  version: 1.5.1
  description: Closed Telegram Mini App and probe API. Unknown values are null, never zero.
servers:
  - url: /api/v1
...
paths:
  /sessions:
    post:
      tags: [Session]
      summary: Exchange verified Telegram initData for a short web session
      operationId: createSession
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [init_data]
              properties:
                init_data: { type: string, minLength: 1, maxLength: 8192 }
      responses:
        "204":
          description: Session cookie created
          headers:
            Set-Cookie: { schema: { type: string } }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
  /sessions/current:
    get:
      tags: [Session]
      summary: Who am I — role of the current web session
      operationId: getSession
      security: [{ webSession: [] }]
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionInfo" }
        "401": { $ref: "#/components/responses/Problem" }
    delete:
      tags: [Session]
      summary: Revoke current web session
      operationId: deleteSession
      security: [{ webSession: [] }]
      responses:
        "204": { description: Session revoked }
  /status:
    get:
      tags: [Status]
      summary: Current member-safe system status
      operationId: getStatus
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
      security: [{ webSession: [] }]
      responses:
        "200":
          description: Current snapshot
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StatusResponse" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /servers/{serverId}:
    get:
      tags: [Status]
      summary: Member-safe server details
      operationId: getServer
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
      security: [{ webSession: [] }]
      parameters:
        - $ref: "#/components/parameters/ServerId"
      responses:
        "200":
          description: Server details
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServerDetail" }
        "404": { $ref: "#/components/responses/Problem" }
...
  /events:
    get:
      tags: [Status]
      summary: Paginated member-safe event history
      operationId: getEvents
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
      security: [{ webSession: [] }]
      parameters:
        - name: filter
          in: query
          schema: { type: string, enum: [all, problems, notes], default: all }
        - name: cursor
          in: query
          schema: { type: string, maxLength: 512 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200":
          description: Event page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventPage" }
  /help:
    get:
      tags: [Status]
      summary: Localized help content keys and contact action
      operationId: getHelp
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
      security: [{ webSession: [] }]
      responses:
        "200":
          description: Help model
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [step_keys, contact_available]
                properties:
                  step_keys:
                    type: array
                    minItems: 4
                    maxItems: 4
                    items: { type: string, pattern: "^help\\." }
                  contact_available: { type: boolean }
                  contact_url: { type: [string, "null"], format: uri, description: Telegram deep link for "Contact administrator"; null when contact_available is false }
  /admin/servers/{serverId}:
    get:
      tags: [Admin]
      summary: Admin-only diagnostic projection
      operationId: getAdminServer
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
      security: [{ webSession: [] }]
      parameters:
        - $ref: "#/components/parameters/ServerId"
      responses:
        "200":
          description: Diagnostics
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminServerDetail" }
        "403": { $ref: "#/components/responses/Problem" }
  /admin/probes:
    get:
      tags: [Admin]
      summary: Probe coverage and freshness
      operationId: getAdminProbes
      security: [{ webSession: [] }]
      responses:
        "200":
          description: Probe list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProbeSummary" }
  /admin/overview:
    get:
      tags: [Admin]
      summary: Attention list and doctor summary for the administrator dashboard
      operationId: getAdminOverview
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
      security: [{ webSession: [] }]
      responses:
        "200":
          description: Overview
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminOverview" }
        "403": { $ref: "#/components/responses/Problem" }
  /admin/probe-enrollments:
    post:
      tags: [Admin]
      summary: Create a single-use ten-minute enrollment code
      operationId: createProbeEnrollment
      security: [{ webSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EnrollmentRequest" }
      responses:
        "201":
          description: Enrollment code, shown once
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [code, expires_at]
                properties:
                  code: { type: string, minLength: 20 }
                  expires_at: { type: string, format: date-time }
        "413": { $ref: "#/components/responses/Problem" }
  /admin/probes/{probeId}/revoke:
    post:
      tags: [Admin]
      summary: Revoke one probe token
      operationId: revokeProbe
      security: [{ webSession: [] }]
      parameters:
        - name: probeId
          in: path
          required: true
          schema: { $ref: "#/components/schemas/PublicId" }
      responses:
        "204": { description: Probe revoked }
        "404": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
  /admin/note:
    put:
      tags: [Admin]
      summary: Create or replace the active administrator note
      operationId: putAdminNote
      security: [{ webSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminNoteInput" }
      responses:
        "200":
          description: Saved note
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminNote" }
        "404": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
    delete:
      tags: [Admin]
      summary: Remove active administrator note
      operationId: deleteAdminNote
      security: [{ webSession: [] }]
      responses:
        "204": { description: Note removed }
  /probe/enroll:
    post:
      tags: [Probes]
      summary: Exchange enrollment code for a personal token
      operationId: enrollProbe
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [code, agent_version, schema_version]
              properties:
                code: { type: string }
                agent_version: { type: string, maxLength: 64 }
                schema_version: { type: integer, const: 1 }
      responses:
        "201":
          description: Personal token and allowed configuration
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [token, probe, config]
                properties:
                  token: { type: string, minLength: 32 }
                  probe: { $ref: "#/components/schemas/ProbeSummary" }
                  config: { $ref: "#/components/schemas/ProbeConfig" }
        "401": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
  /probe/config:
    get:
      tags: [Probes]
      summary: Get capability-scoped probe configuration
      operationId: getProbeConfig
      security: [{ probeToken: [] }]
      responses:
        "200":
          description: Configuration
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProbeConfig" }
  /probe/reports:
    post:
      tags: [Probes]
      summary: Submit an idempotent normalized probe report
      operationId: createProbeReport
      security: [{ probeToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProbeReport" }
      responses:
        "202":
          description: Accepted or previously accepted
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [report_id, accepted, duplicate]
                properties:
                  report_id: { type: string, format: uuid }
                  accepted: { type: boolean }
                  duplicate: { type: boolean }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
  /analytics/events:batch:
    post:
      tags: [Analytics]
      summary: Submit privacy-safe allowlisted product events
      operationId: createAnalyticsBatch
      security: [{ webSession: [] }, { probeToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [events]
              properties:
                events:
                  type: array
                  minItems: 1
                  maxItems: 50
                  items: { $ref: "#/components/schemas/ProductEvent" }
      responses:
        "202": { description: Batch accepted idempotently }
        "400": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    webSession:
      type: apiKey
      in: cookie
      name: vpnpulse_session
    probeToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: >-
        language of server names, the administrator note and doctor hints: the first supported tag
        (ru or en) wins, otherwise the installation default. Everything else is an i18n key the
        client renders itself.
      schema: { type: string, example: "ru-RU,ru;q=0.9,en;q=0.5" }
    ServerId:
      name: serverId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/PublicId" }
  responses:
    Problem:
      description: RFC 9457 problem details
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    PublicId:
      type: string
      pattern: "^[a-z][a-z0-9-]{1,63}$"
    State:
      type: string
      enum: [operational, degraded, unavailable, unknown]
...
    StatusResponse:
      type: object
      additionalProperties: false
      required: [state, freshness, coverage, servers]
      properties:
        state: { $ref: "#/components/schemas/State" }
        freshness: { $ref: "#/components/schemas/Freshness" }
        coverage: { type: number, minimum: 0, maximum: 1 }
        mode:
          description: live monitoring or fixture-backed demo; the UI shows a permanent "Demo data" mark for demo
          type: string
          enum: [live, demo]
          default: live
        observing_since: { type: [string, "null"], format: date-time, description: first observation ever; used by the empty events state }
        recommended_server_id: { anyOf: [{ $ref: "#/components/schemas/PublicId" }, { type: "null" }] }
        note: { anyOf: [{ $ref: "#/components/schemas/AdminNote" }, { type: "null" }] }
        servers:
          description: empty on a clean install (no servers connected yet)
          type: array
          minItems: 0
          items: { $ref: "#/components/schemas/ServerCard" }
        sources:
          description: >-
            check sources that exist right now. A source appears after its probe's first accepted
            report (the initialising event) and disappears when the probe is revoked; the UI shows
            only listed sources in the legend, rows and checks. Missing sources are not an incident.
          type: array
          maxItems: 3
          items: { $ref: "#/components/schemas/SourceAvailability" }
...
    ServerDetail:
      allOf:
        - $ref: "#/components/schemas/ServerCard"
        - type: object
          required: [protocols, checks]
          properties:
            protocols:
              type: array
              items: { $ref: "#/components/schemas/ProtocolSummary" }
            checks:
              description: only sources listed in /status.sources; empty when no probe has reported yet
              type: array
              items: { $ref: "#/components/schemas/EvidenceSummary" }
            uptime_7d: { type: [number, "null"], minimum: 0, maximum: 1 }
            coverage_7d: { type: [number, "null"], minimum: 0, maximum: 1 }
            components:
              description: member-safe software list for the "Software" group (no hosts, ports or paths)
              type: array
              maxItems: 12
              items: { $ref: "#/components/schemas/SoftwareComponent" }
            resources: { anyOf: [{ $ref: "#/components/schemas/ResourceSnapshot" }, { type: "null" }] }
            profiles: { anyOf: [{ $ref: "#/components/schemas/ProfileStats" }, { type: "null" }] }
            service_checks:
              description: binary collector checks for the "Server" group
              type: array
              maxItems: 8
              items: { $ref: "#/components/schemas/ServiceCheck" }
...
    EventPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Event" } }
        next_cursor: { type: [string, "null"] }
    AdminServerDetail:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/ServerDetail"
        - type: object
          required: [attention, diagnostics]
          properties:
            attention: { type: array, items: { type: string, maxLength: 80 } }
            attention_items:
              description: structured attention list for the admin dashboard (severity-ordered)
              type: array
              maxItems: 50
              items: { $ref: "#/components/schemas/AttentionItem" }
            diagnostics: { type: object, additionalProperties: true }
    SessionInfo:
      type: object
      additionalProperties: false
      required: [role, expires_at]
      properties:
        role: { type: string, enum: [member, admin] }
        expires_at: { type: string, format: date-time }
...
    EnrollmentRequest:
      type: object
      additionalProperties: false
      required: [kind, capabilities]
      properties:
        kind: { type: string, enum: [pc, android, abroad, watchdog] }
        capabilities:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string, enum: [report_pc, report_mobile, report_abroad, read_watchdog] }
...